+++ src/kava/scan/kava8.py
@@ -1,0 +1,4 @@
+from .base import KavaBase
+
+class Kava8B(KavaBase):
+    pass
